requirements.txt
Flask~=3.0.0
Flask-Cors==4.0.0
Pillow~=10.1.0
opencv-python~=4.8.1.78
numpy~=1.26.2
torch~=2.1.1
torchvision~=0.16.1
scikit-learn~=1.3.2
tqdm~=4.66.1
scikit-learn~=1.5.0
pgvector~=0.2.5
psycopg-binary~=3.1.19
pokemontcgsdk~=3.4.0